PoC/compliledger/backend/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.7
pydantic==2.5.3
google-generativeai==0.3.1
py-algorand-sdk==2.5.0
httpx==0.26.0
python-dotenv==1.0.0
pyteal==0.10.1
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
